renamed Main.js to Main.jsx
