--- src/layout/main/Main.js
@@ -1,19 +1,0 @@
-import styled from 'styled-components'
-import { MainContainer } from '../../components/stupidComponents/container/MainContainer.styled'
-
-export const Main = () => {
-    return (
-      <StyledMain className='main'>
-          <MainContainer direction={'column'} gap={'40px'} justify={'center'}>
-            fgjhfbkgfj
-          </MainContainer>
-      </StyledMain>
-    )
-  }
-  
-  const StyledMain = styled.main`
-      width: 100%;
-      border: 1px solid red;
-      padding-top: 30px;
-      padding-bottom: 30px;
-  `
